refac: pallet subspace
More organized dependency management in pallet subspace
