+++ pallets/subspace/src/network/mod.rs
@@ -1,0 +1,4 @@
+pub mod module;
+pub mod registration;
+pub mod staking;
+pub mod subnet;
