+++ pallets/subspace/src/network/registration.rs
@@ -1,5 +1,4 @@
-use super::*;
-use crate::{module::ModuleChangeset, subnet::SubnetChangeset};
+use crate::*;
 // use frame_support::storage::with_storage_layer;
 
 use frame_support::{
