+++ pallets/subspace/src/network/staking.rs
@@ -1,4 +1,4 @@
-use super::*;
+use crate::*;
 
 use sp_arithmetic::per_things::Percent;
 use sp_runtime::DispatchError;
