+++ pallets/subspace/src/params/global.rs
@@ -1,4 +1,5 @@
-use super::*;
+use crate::*;
+
 use frame_support::pallet_prelude::{DispatchResult, MaxEncodedLen};
 use sp_core::Get;
 use sp_runtime::DispatchError;
