+++ pallets/subspace/src/params/mod.rs
@@ -1,0 +1,2 @@
+pub mod global;
+pub mod params;
